:recycle: Move dataset integration activation code to separate function.
Why? For easier reuse during dataset update, as activation needs to be refreshed if url or headers are changed.
