+++ api/src/models/dataset-integrations/index.ts
@@ -1,0 +1,1 @@
+export { activate } from "./activate"
